Create bootstrap.ts
